refactor(app): split browser transport runtime
